Refactor get datashuttle path functions to canonical_folders.
